💻 Projects/skillbridge/pyproject.toml
[project]
name = "skillbridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "streamlit[standard]>=1.45.1",
]
